sub_project/sub_backend/backend-py/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
PyJWT==2.8.0
redis==5.0.4
minio==7.2.9
pydantic==2.6.4
SQLAlchemy==2.0.29
PyMySQL==1.1.1
openpyxl==3.1.2
